change server port to short

+++ 1/tcpEchoClient.c
@@ -27,7 +27,7 @@ int main(int argc, char* argv[])
 	int s;
 	int len;
 	struct packet packet_reg;
-	int SERVER_PORT;
+	short SERVER_PORT;
 
 	if(argc == 3){
 		host = argv[1];
